server/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.2
django-cors-headers==3.6.0
psycopg2-binary>=2.8
gunicorn==20.0.4
black==20.8b1
Pillow==8.1.0
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.1
django-filter==2.4.0
django-downloadview==2.1.1
django-rest-framework-condition==0.1.1

# Dependencies for SocketIO
enum-compat==0.0.3
eventlet==0.30.0
greenlet==1.1.2
python-engineio
python-socketio
pytz==2018.7
six==1.10.0
# End dependencies for SocketIO
